requirements.txt
flask
flask-cors
google-generativeai
tensorflow
Pillow
scikit-learn